refactor: Rename methods/ to methodlevel/
